+++ CHANGES.md
@@ -9,6 +9,7 @@
 ##### Additions :tada:
 
 - Added `IndexDatatype.fromTypedArray`. [#10350](https://github.com/CesiumGS/cesium/pull/10350)
+- Fixed crash for zero-area `region` bounding volumes in a 3D Tileset.
 
 ##### Fixes :wrench: